modified:   .gitignore
	modified:   app/Http/Controllers/CS/CsController.php
	modified:   app/Http/Controllers/CS/ReportController.php
	modified:   app/Http/Controllers/HomeController.php
	modified:   app/Http/Controllers/MonitoringController.php
	modified:   app/Http/Controllers/Supervisor/CrudCsController.php
	modified:   app/Http/Controllers/Supervisor/CrudRoomController.php
	modified:   app/Http/Controllers/Supervisor/ReportController.php
	modified:   app/Http/Controllers/Supervisor/ScheduleController.php
	modified:   app/Http/Controllers/Supervisor/SupervisorController.php
	modified:   app/Libraries/Fungsi.php
	modified:   public/assets/css/custom.css
	new file:   public/assets/css/magnific-popup.css
	modified:   public/assets/css/partials/_general.scss
	modified:   public/assets/css/style.css
	new file:   public/assets/js/jquery.magnific-popup.min.js
	new file:   public/assets/js/magnific-popup.min.js
	modified:   public/assets/scss/partials/_general.scss
	modified:   resources/views/layouts/footer.blade.php
	modified:   resources/views/layouts/header.blade.php
	modified:   resources/views/layouts/template/supervisor-sidebar.blade.php
	modified:   resources/views/supervisor/schedule/edit.blade.php
	modified:   resources/views/supervisor/schedule/index.blade.php
	modified:   routes/web.php
	new file:   storage.zip

+++ app/Http/Controllers/Supervisor/CrudCsController.php
@@ -4,12 +4,13 @@ namespace App\Http\Controllers\Supervisor;
 
 use App\Models\Role;
 use App\Models\User;
+use App\Models\Schedule;
 use App\Libraries\Fungsi;
 use Illuminate\Support\Str;
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rule;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Hash;
 
 class CrudCsController extends Controller
 {
@@ -88,6 +89,10 @@ class CrudCsController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
+        if (count(Schedule::where('cs_id', $id)->get()) > 0) {
+            Fungsi::sweetalert('Akun Cleaning Service tidak dapat dihapus karena sudah digunakan', 'error', 'Gagal!');
+            return redirect()->route('supervisor.cs.data');
+        }
         if ($user["avatar"] != 'assets/img/user/default.jpg') {
             unlink(storage_path('app/public/' . $user["avatar"]));
         }
